pyproject.toml
[project]
name = "appium-automation"
version = "1.0.0"
description = "Scalable Appium testing framework for iOS and Android"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Appium-Python-Client>=3.1.0",
    "pytest>=7.0.0",
    "pytest-html>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-html>=3.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "android: marks tests as Android tests",
    "ios: marks tests as iOS tests",
    "smoke: marks tests as smoke tests",
    "regression: marks tests as regression tests",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

